ProtypePruebas/Backend/Backend/requirements.txt
flask==2.3.2
flask-jwt-extended==4.5.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
pyjwt==2.8.0
cryptography==42.0.2